+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - No need to have the overhead of pre-registering every e-mail address as a 'notify entity'
   - Attachment, template and HTML behaviour identical to existing email integration
   - Emails using `smtp` support `Importance`,`Priority`,`X-Priority` and `X-MSMail-Priority` headers. Values are mapped automatically from the Supernotify priority.
+  - `Message-Id` in email header ties back to notification ID and delivery name
 - `delivery_selection` of `fixed` overrides priority and condition filtering at delivery level, if selection is fixed, only `enabled` at transport or delivery level will override your command.
 
 ## 1.16.6